requirements_ci.txt
bandit
coverage
mock>=1.0.1
flake8>=2.1.0
coveralls
django-model-utils>=2.0

# Additional test requirements go here
setuptools
wheel
twine
pbr